\docType{package}
\name{ghit}
\alias{ghit}
\alias{ghit-package}
\title{Lightweight GitHub Package Installer}
\description{ghit provides a single function, \code{\link{install_github}}, which allows users to install R packages hosted on GitHub without needing to install or load the heavy dependencies required by devtools. ghit provides a drop-in replacement that provides (almost) identical functionality.}
