% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghit-package.R
\docType{package}
\name{ghit}
\alias{ghit}
\alias{ghit-package}
\title{Lightweight GitHub Package Installer}
\description{
ghit provides two functions, \code{\link{install_github}} and \code{\link{install_bitbucket}}, which allow users to install R packages hosted on GitHub or Bitbucket without needing to install or load the heavy dependencies required by devtools. ghit provides a drop-in replacement that provides (almost) identical functionality.
}

